add statistics

See merge request s13-fintech-finance-sub1/S13P21A407!86

+++ pocketc_backend/src/main/java/com/ssafy/pocketc_backend/domain/report/dto/response/MonthlySavingScoreDto.java
@@ -4,9 +4,9 @@ import java.time.LocalDate;
 
 public record MonthlySavingScoreDto(
         LocalDate month,
-        Integer savingScore
+        Double savingScore
 ) {
-    public static MonthlySavingScoreDto of(LocalDate month, Integer savingScore) {
+    public static MonthlySavingScoreDto of(LocalDate month, Double savingScore) {
         return new MonthlySavingScoreDto(month, savingScore);
     }
 }
